fix(outlook): stop truncating fractional work hours, retry work-week lookup after transient failure
Three review findings on find_meeting_slots' new EWS work-week lookup:

- WorkWeekInfo.StartHour/EndHour truncated via (int) cast, so a mailbox
  configured for e.g. 08:30-17:30 silently became 08:00-17:00 - offering
  a slot before the real work day starts and dropping the valid
  17:00-17:30 slot. Now kept as fractional hours end-to-end
  (MeetingSlots.Rank takes double, not int; display formatted as HH:mm).
- ResolveWorkWeekAsync latched _workWeekResolved=true before the lookup
  even ran, so one transient EWS/network failure permanently disabled
  the real work-week lookup for the rest of the Outlook session (which
  can run for days), unlike OutlookEws.CachedUrl - which only caches on
  success and retries after a failure, as this code's own comment
  claimed it mirrored but didn't. Now only latches on success.
- entry.ts's start_hour/end_hour descriptions still said "(default 9)"/
  "(default 18)" despite the tool now defaulting to the mailbox's real
  resolved work hours.

Verified: OfficeAi.Shared.Tests 165/165, OutlookAiAddIn MSBuild clean,
tsc --noEmit clean in all four add-ins.

+++ OutlookAiAddIn/OutlookEws.cs
@@ -91,8 +91,8 @@ namespace OutlookAiAddIn
         internal struct WorkWeekInfo
         {
             public HashSet<DayOfWeek> Days;
-            public int StartHour;
-            public int EndHour;
+            public double StartHour;
+            public double EndHour;
         }
 
         // GetUserAvailability's WorkingHours is EWS's documented, server-side
@@ -151,8 +151,12 @@ namespace OutlookAiAddIn
                 return new WorkWeekInfo
                 {
                     Days = days,
-                    StartHour = (int)a.WorkingHours.StartTime.TotalHours,
-                    EndHour = (int)a.WorkingHours.EndTime.TotalHours,
+                    // Kept as fractional hours, not truncated to an int: a
+                    // mailbox configured for e.g. 08:30-17:30 would otherwise
+                    // silently become 08:00-17:00 (offering a slot before the
+                    // real start, and dropping the valid 17:00-17:30 slot).
+                    StartHour = a.WorkingHours.StartTime.TotalHours,
+                    EndHour = a.WorkingHours.EndTime.TotalHours,
                 };
             }
             return null;
